Moved Node and Edge interfaces into single file with Automata
