--- frontend/src/automata/Edge.ts
@@ -1,5 +1,0 @@
-export interface IEdge {
-    from: number,
-    to: number,
-    label: string
-}
