--- frontend/src/automata/Node.ts
@@ -1,4 +1,0 @@
-export interface INode {
-    id: number,
-    isTerminal: boolean
-}
